+++ ext/node/lib.rs
@@ -13,6 +13,7 @@ use deno_semver::npm::NpmPackageNvReference;
 use deno_semver::npm::NpmPackageReq;
 use deno_semver::npm::NpmPackageReqReference;
 use once_cell::sync::Lazy;
+use resolution::NodeResolutionMode;
 use std::collections::HashSet;
 use std::io;
 use std::path::Path;
@@ -27,7 +28,6 @@ mod package_json;
 mod path;
 mod polyfill;
 mod resolution;
-mod resolver;
 
 pub use package_json::PackageJson;
 pub use path::PathClean;
@@ -35,22 +35,14 @@ pub use polyfill::is_builtin_node_module;
 pub use polyfill::resolve_builtin_node_module;
 pub use polyfill::NodeModulePolyfill;
 pub use polyfill::SUPPORTED_BUILTIN_NODE_MODULES;
-pub use resolution::get_closest_package_json;
-pub use resolution::get_package_scope_config;
-pub use resolution::legacy_main_resolve;
-pub use resolution::package_exports_resolve;
-pub use resolution::package_imports_resolve;
-pub use resolution::package_resolve;
-pub use resolution::path_to_declaration_path;
 pub use resolution::NodeModuleKind;
-pub use resolution::NodeResolutionMode;
-pub use resolution::DEFAULT_CONDITIONS;
-pub use resolver::NodeResolution;
-pub use resolver::NodeResolver;
+pub use resolution::NodeResolution;
+pub use resolution::NodeResolver;
 
 pub trait NodeEnv {
   type P: NodePermissions;
   type Fs: NodeFs;
+  type NpmResolver: NpmResolver;
 }
 
 pub trait NodePermissions {
@@ -71,24 +63,26 @@ pub struct NodeFsMetadata {
   pub is_dir: bool,
 }
 
-pub trait NodeFs {
-  fn current_dir() -> io::Result<PathBuf>;
-  fn metadata<P: AsRef<Path>>(path: P) -> io::Result<NodeFsMetadata>;
-  fn is_file<P: AsRef<Path>>(path: P) -> bool;
-  fn is_dir<P: AsRef<Path>>(path: P) -> bool;
-  fn exists<P: AsRef<Path>>(path: P) -> bool;
-  fn read_to_string<P: AsRef<Path>>(path: P) -> io::Result<String>;
-  fn canonicalize<P: AsRef<Path>>(path: P) -> io::Result<PathBuf>;
+pub trait NodeFs: Clone {
+  fn current_dir(&self) -> io::Result<PathBuf>;
+  fn metadata(&self, path: impl AsRef<Path>) -> io::Result<NodeFsMetadata>;
+  fn is_file(&self, path: impl AsRef<Path>) -> bool;
+  fn is_dir(&self, path: impl AsRef<Path>) -> bool;
+  fn exists(&self, path: impl AsRef<Path>) -> bool;
+  fn read_to_string(&self, path: impl AsRef<Path>) -> io::Result<String>;
+  fn canonicalize(&self, path: impl AsRef<Path>) -> io::Result<PathBuf>;
 }
 
+#[derive(Clone)]
 pub struct RealFs;
+
 impl NodeFs for RealFs {
-  fn current_dir() -> io::Result<PathBuf> {
+  fn current_dir(&self) -> io::Result<PathBuf> {
     #[allow(clippy::disallowed_methods)]
     std::env::current_dir()
   }
 
-  fn metadata<P: AsRef<Path>>(path: P) -> io::Result<NodeFsMetadata> {
+  fn metadata(&self, path: impl AsRef<Path>) -> io::Result<NodeFsMetadata> {
     #[allow(clippy::disallowed_methods)]
     std::fs::metadata(path).map(|metadata| {
       // on most systems, calling is_file() and is_dir() is cheap
@@ -100,29 +94,29 @@ impl NodeFs for RealFs {
     })
   }
 
-  fn exists<P: AsRef<Path>>(path: P) -> bool {
+  fn exists(&self, path: impl AsRef<Path>) -> bool {
     #[allow(clippy::disallowed_methods)]
     std::fs::metadata(path).is_ok()
   }
 
-  fn is_file<P: AsRef<Path>>(path: P) -> bool {
+  fn is_file(&self, path: impl AsRef<Path>) -> bool {
     #[allow(clippy::disallowed_methods)]
     std::fs::metadata(path)
       .map(|m| m.is_file())
       .unwrap_or(false)
   }
 
-  fn is_dir<P: AsRef<Path>>(path: P) -> bool {
+  fn is_dir(&self, path: impl AsRef<Path>) -> bool {
     #[allow(clippy::disallowed_methods)]
     std::fs::metadata(path).map(|m| m.is_dir()).unwrap_or(false)
   }
 
-  fn read_to_string<P: AsRef<Path>>(path: P) -> io::Result<String> {
+  fn read_to_string(&self, path: impl AsRef<Path>) -> io::Result<String> {
     #[allow(clippy::disallowed_methods)]
     std::fs::read_to_string(path)
   }
 
-  fn canonicalize<P: AsRef<Path>>(path: P) -> io::Result<PathBuf> {
+  fn canonicalize(&self, path: impl AsRef<Path>) -> io::Result<PathBuf> {
     #[allow(clippy::disallowed_methods)]
     std::path::Path::canonicalize(path.as_ref())
   }
@@ -582,12 +576,18 @@ deno_core::extension!(deno_node,
     "zlib.ts",
   ],
   options = {
+    maybe_node_resolver: Option<NodeResolver<Env::Fs, Env::NpmResolver>>,
     maybe_npm_resolver: Option<Rc<dyn NpmResolver>>,
+    fs: Env::Fs,
   },
   state = |state, options| {
     if let Some(npm_resolver) = options.maybe_npm_resolver {
       state.put(npm_resolver);
     }
+    if let Some(node_resolver) = options.maybe_node_resolver {
+      state.put(node_resolver);
+    }
+    state.put(options.fs);
   },
 );
