+++ ext/node/ops/require.rs
@@ -20,6 +20,7 @@ use crate::NodeFs;
 use crate::NodeModuleKind;
 use crate::NodePermissions;
 use crate::NodeResolutionMode;
+use crate::NodeResolver;
 use crate::NpmResolver;
 use crate::PackageJson;
 
@@ -96,10 +97,11 @@ pub fn op_require_node_module_paths<Env>(
 where
   Env: NodeEnv + 'static,
 {
+  let fs = state.borrow::<Env::Fs>().clone();
   // Guarantee that "from" is absolute.
   let from = deno_core::resolve_path(
     &from,
-    &(Env::Fs::current_dir()).context("Unable to get CWD")?,
+    &(fs.current_dir()).context("Unable to get CWD")?,
   )
   .unwrap()
   .to_file_path()
@@ -264,7 +266,8 @@ where
 {
   let path = PathBuf::from(path);
   ensure_read_permission::<Env::P>(state, &path)?;
-  if let Ok(metadata) = Env::Fs::metadata(&path) {
+  let fs = state.borrow::<Env::Fs>().clone();
+  if let Ok(metadata) = fs.metadata(&path) {
     if metadata.is_file {
       return Ok(0);
     } else {
@@ -285,7 +288,8 @@ where
 {
   let path = PathBuf::from(request);
   ensure_read_permission::<Env::P>(state, &path)?;
-  let mut canonicalized_path = Env::Fs::canonicalize(&path)?;
+  let fs = state.borrow::<Env::Fs>().clone();
+  let mut canonicalized_path = fs.canonicalize(&path)?;
   if cfg!(windows) {
     canonicalized_path = PathBuf::from(
       canonicalized_path
@@ -353,7 +357,8 @@ where
 
   if let Some(parent_id) = maybe_parent_id {
     if parent_id == "<repl>" || parent_id == "internal/preload" {
-      if let Ok(cwd) = Env::Fs::current_dir() {
+      let fs = state.borrow::<Env::Fs>().clone();
+      if let Ok(cwd) = fs.current_dir() {
         ensure_read_permission::<Env::P>(state, &cwd)?;
         return Ok(Some(cwd.to_string_lossy().to_string()));
       }
@@ -375,14 +380,16 @@ where
     return Ok(None);
   }
 
-  let resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
+  let node_resolver = state
+    .borrow::<Rc<NodeResolver<Env::Fs, Env::NpmResolver>>>()
+    .clone();
   let permissions = state.borrow_mut::<Env::P>();
-  let pkg = resolution::get_package_scope_config::<Env::Fs>(
-    &Url::from_file_path(parent_path.unwrap()).unwrap(),
-    &*resolver,
-    permissions,
-  )
-  .ok();
+  let pkg = node_resolver
+    .get_package_scope_config(
+      &Url::from_file_path(parent_path.unwrap()).unwrap(),
+      permissions,
+    )
+    .ok();
   if pkg.is_none() {
     return Ok(None);
   }
@@ -408,18 +415,18 @@ where
 
   let referrer = deno_core::url::Url::from_file_path(&pkg.path).unwrap();
   if let Some(exports) = &pkg.exports {
-    resolution::package_exports_resolve::<Env::Fs>(
-      &pkg.path,
-      expansion,
-      exports,
-      &referrer,
-      NodeModuleKind::Cjs,
-      resolution::REQUIRE_CONDITIONS,
-      NodeResolutionMode::Execution,
-      &*resolver,
-      permissions,
-    )
-    .map(|r| Some(r.to_string_lossy().to_string()))
+    node_resolver
+      .package_exports_resolve(
+        &pkg.path,
+        expansion,
+        exports,
+        &referrer,
+        NodeModuleKind::Cjs,
+        resolution::REQUIRE_CONDITIONS,
+        NodeResolutionMode::Execution,
+        permissions,
+      )
+      .map(|r| Some(r.to_string_lossy().to_string()))
   } else {
     Ok(None)
   }
@@ -435,7 +442,8 @@ where
 {
   let file_path = PathBuf::from(file_path);
   ensure_read_permission::<Env::P>(state, &file_path)?;
-  Ok(Env::Fs::read_to_string(file_path)?)
+  let fs = state.borrow::<Env::Fs>().clone();
+  Ok(fs.read_to_string(file_path)?)
 }
 
 #[op]
@@ -462,10 +470,14 @@ fn op_require_resolve_exports<Env>(
 where
   Env: NodeEnv + 'static,
 {
-  let resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
+  let fs = state.borrow::<Env::Fs>().clone();
+  let npm_resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
+  let node_resolver = state
+    .borrow::<Rc<NodeResolver<Env::Fs, Env::NpmResolver>>>()
+    .clone();
   let permissions = state.borrow_mut::<Env::P>();
 
-  let pkg_path = if resolver
+  let pkg_path = if npm_resolver
     .in_npm_package_at_path(&PathBuf::from(&modules_path))
     && !uses_local_node_modules_dir
   {
@@ -473,32 +485,33 @@ where
   } else {
     let orignal = modules_path.clone();
     let mod_dir = path_resolve(vec![modules_path, name]);
-    if Env::Fs::is_dir(&mod_dir) {
+    if fs.is_dir(&mod_dir) {
       mod_dir
     } else {
       orignal
     }
   };
-  let pkg = PackageJson::load::<Env::Fs>(
-    &*resolver,
+  let pkg = PackageJson::load(
+    &fs,
+    &*npm_resolver,
     permissions,
     PathBuf::from(&pkg_path).join("package.json"),
   )?;
 
   if let Some(exports) = &pkg.exports {
     let referrer = Url::from_file_path(parent_path).unwrap();
-    resolution::package_exports_resolve::<Env::Fs>(
-      &pkg.path,
-      format!(".{expansion}"),
-      exports,
-      &referrer,
-      NodeModuleKind::Cjs,
-      resolution::REQUIRE_CONDITIONS,
-      NodeResolutionMode::Execution,
-      &*resolver,
-      permissions,
-    )
-    .map(|r| Some(r.to_string_lossy().to_string()))
+    node_resolver
+      .package_exports_resolve(
+        &pkg.path,
+        format!(".{expansion}"),
+        exports,
+        &referrer,
+        NodeModuleKind::Cjs,
+        resolution::REQUIRE_CONDITIONS,
+        NodeResolutionMode::Execution,
+        permissions,
+      )
+      .map(|r| Some(r.to_string_lossy().to_string()))
   } else {
     Ok(None)
   }
@@ -516,11 +529,12 @@ where
     state,
     PathBuf::from(&filename).parent().unwrap(),
   )?;
-  let resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
+  let node_resolver = state
+    .borrow::<Rc<NodeResolver<Env::Fs, Env::NpmResolver>>>()
+    .clone();
   let permissions = state.borrow_mut::<Env::P>();
-  resolution::get_closest_package_json::<Env::Fs>(
+  node_resolver.get_closest_package_json(
     &Url::from_file_path(filename).unwrap(),
-    &*resolver,
     permissions,
   )
 }
@@ -533,10 +547,11 @@ fn op_require_read_package_scope<Env>(
 where
   Env: NodeEnv + 'static,
 {
+  let fs = state.borrow::<Env::Fs>().clone();
   let resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
   let permissions = state.borrow_mut::<Env::P>();
   let package_json_path = PathBuf::from(package_json_path);
-  PackageJson::load::<Env::Fs>(&*resolver, permissions, package_json_path).ok()
+  PackageJson::load(&fs, &*resolver, permissions, package_json_path).ok()
 }
 
 #[op]
@@ -550,9 +565,14 @@ where
 {
   let parent_path = PathBuf::from(&parent_filename);
   ensure_read_permission::<Env::P>(state, &parent_path)?;
+  let fs = state.borrow::<Env::Fs>().clone();
   let resolver = state.borrow::<Rc<dyn NpmResolver>>().clone();
+  let node_resolver = state
+    .borrow::<Rc<NodeResolver<Env::Fs, Env::NpmResolver>>>()
+    .clone();
   let permissions = state.borrow_mut::<Env::P>();
-  let pkg = PackageJson::load::<Env::Fs>(
+  let pkg = PackageJson::load(
+    &fs,
     &*resolver,
     permissions,
     parent_path.join("package.json"),
@@ -561,18 +581,16 @@ where
   if pkg.imports.is_some() {
     let referrer =
       deno_core::url::Url::from_file_path(&parent_filename).unwrap();
-    let r = resolution::package_imports_resolve::<Env::Fs>(
-      &request,
-      &referrer,
-      NodeModuleKind::Cjs,
-      resolution::REQUIRE_CONDITIONS,
-      NodeResolutionMode::Execution,
-      &*resolver,
-      permissions,
-    )
-    .map(|r| Some(Url::from_file_path(r).unwrap().to_string()));
-    state.put(resolver);
-    r
+    node_resolver
+      .package_imports_resolve(
+        &request,
+        &referrer,
+        NodeModuleKind::Cjs,
+        resolution::REQUIRE_CONDITIONS,
+        NodeResolutionMode::Execution,
+        permissions,
+      )
+      .map(|r| Some(Url::from_file_path(r).unwrap().to_string()))
   } else {
     Ok(None)
   }
